requirements.txt
pyannote.audio
numpy~=2.0.2
scikit-learn~=1.6.0
tqdm~=4.67.1

hdbscan~=0.8.40
python-dotenv~=1.0.1
librosa~=0.10.2.post1